+++ tasks/handler/worker.go
@@ -320,8 +320,16 @@ func HandleChannelEvent(ctx context.Context, db *sqlx.DB, rp *redis.Pool, eventT
 
 	modelContact := contacts[0]
 
+	if models.ContactSeenEvents[eventType] {
+		err = modelContact.UpdateLastSeenOn(ctx, db, event.OccurredOn())
+		if err != nil {
+			return nil, errors.Wrap(err, "error updating contact last_seen_on")
+		}
+	}
+
 	// do we have associated trigger?
 	var trigger *models.Trigger
+
 	switch eventType {
 
 	case models.NewConversationEventType:
@@ -336,7 +344,7 @@ func HandleChannelEvent(ctx context.Context, db *sqlx.DB, rp *redis.Pool, eventT
 	case models.MOCallEventType:
 		trigger = models.FindMatchingMOCallTrigger(oa, modelContact)
 
-	case models.WelcomeMessateEventType:
+	case models.WelcomeMessageEventType:
 		trigger = nil
 
 	default:
@@ -449,11 +457,19 @@ func handleStopEvent(ctx context.Context, db *sqlx.DB, rp *redis.Pool, event *St
 	if err != nil {
 		return errors.Wrapf(err, "unable to start transaction for stopping contact")
 	}
+
 	err = models.StopContact(ctx, tx, event.OrgID, event.ContactID)
 	if err != nil {
 		tx.Rollback()
 		return err
 	}
+
+	err = models.UpdateContactLastSeenOn(ctx, tx, event.ContactID, event.OccurredOn)
+	if err != nil {
+		tx.Rollback()
+		return err
+	}
+
 	err = tx.Commit()
 	if err != nil {
 		return errors.Wrapf(err, "unable to commit for contact stop")
@@ -631,10 +647,25 @@ func handleMsgEvent(ctx context.Context, db *sqlx.DB, rp *redis.Pool, event *Msg
 		return nil
 	}
 
-	err = models.UpdateMessage(ctx, db, event.MsgID, models.MsgStatusHandled, models.VisibilityVisible, models.TypeInbox, topupID)
+	// this message didn't trigger and new sessions or resume any existing ones, so handle as inbox
+	err = handleAsInbox(ctx, db, modelContact.ID(), event.MsgID, topupID, event.CreatedOn)
+	if err != nil {
+		return errors.Wrapf(err, "error handling inbox message")
+	}
+	return nil
+}
+
+func handleAsInbox(ctx context.Context, db *sqlx.DB, contactID models.ContactID, msgID flows.MsgID, topupID models.TopupID, createdOn time.Time) error {
+	err := models.UpdateMessage(ctx, db, msgID, models.MsgStatusHandled, models.VisibilityVisible, models.TypeInbox, topupID)
 	if err != nil {
 		return errors.Wrapf(err, "error marking message as handled")
 	}
+
+	err = models.UpdateContactLastSeenOn(ctx, db, contactID, createdOn)
+	if err != nil {
+		return errors.Wrapf(err, "error updating contact last_seen_on")
+	}
+
 	return nil
 }
 
@@ -662,11 +693,13 @@ type MsgEvent struct {
 	Text          string             `json:"text"`
 	Attachments   []utils.Attachment `json:"attachments"`
 	NewContact    bool               `json:"new_contact"`
+	CreatedOn     time.Time          `json:"created_on"`
 }
 
 type StopEvent struct {
-	ContactID models.ContactID `json:"contact_id"`
-	OrgID     models.OrgID     `json:"org_id"`
+	ContactID  models.ContactID `json:"contact_id"`
+	OrgID      models.OrgID     `json:"org_id"`
+	OccurredOn time.Time        `json:"occurred_on"`
 }
 
 // NewTimeoutEvent creates a new event task for the passed in timeout event